// backend/src/ood/enums.ts
export enum Color {
    black = 'black' ,
    white = 'white'
}

export enum PieceType{
    pawns = 'pawns' ,
    king = 'king' ,
    Queen = 'Queen' ,
    Rookes = 'Rookes' ,
    Knight = 'Knight' , 
    Bishop = 'Bishop' 

}

export const gameStatus = {
    Checked : "Checked" ,
    CheckMate : "Check Mate" , 
    GameOver : "GameOver",
    Continue : "Continue"
  }